refactor: move jp extensions in a top level package
