+++ pkg/jp/functions.go
@@ -1,11 +1,11 @@
-package template
+package jp
 
 import (
 	"context"
 
 	jpfunctions "github.com/jmespath-community/go-jmespath/pkg/functions"
-	"github.com/kyverno/kyverno-json/pkg/engine/template/functions"
-	kyvernofunctions "github.com/kyverno/kyverno-json/pkg/engine/template/kyverno"
+	"github.com/kyverno/kyverno-json/pkg/jp/functions"
+	kyvernofunctions "github.com/kyverno/kyverno-json/pkg/jp/kyverno"
 )
 
 func GetFunctions(ctx context.Context) []jpfunctions.FunctionEntry {
