+++ pkg/jp/functions/wildcard.go
