+++ pkg/jp/kyverno/arithmetic.go
