+++ pkg/jp/kyverno/functionentry.go
