+++ pkg/jp/kyverno/functionentry_test.go
