+++ website/jp/main.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 
 	jpfunctions "github.com/jmespath-community/go-jmespath/pkg/functions"
-	"github.com/kyverno/kyverno-json/pkg/engine/template/functions"
-	kyvernofunctions "github.com/kyverno/kyverno-json/pkg/engine/template/kyverno"
+	"github.com/kyverno/kyverno-json/pkg/jp/functions"
+	kyvernofunctions "github.com/kyverno/kyverno-json/pkg/jp/kyverno"
 )
 
 func main() {
